requirements.txt
neo4j==5.28.2
pytest==8.4.1
openai==1.102.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.4
firebase-admin==7.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
